Add more OOP and make packages more meaningful for better code readability
