--- src/main/java/Tasks/Todo.java
@@ -1,19 +1,0 @@
-package Tasks;
-
-/**
- * Todo class with additional [T] for toString
- */
-public class Todo extends Task {
-
-    protected String by;
-
-    public Todo(String description) {
-        super(description);
-    }
-
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-
-
-}
