+++ src/main/java/tasks/Deadline.java
@@ -1,7 +1,12 @@
-package Tasks;
+package tasks;
+
+import static resources.DateHandler.isDate;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
+
+import resources.Constants;
+import resources.Helper;
 
 /**
  * Deadline class with additional field 'by'
@@ -34,5 +39,23 @@ public class Deadline extends Task {
         return this.by;
     }
 
+    /**
+     * respond to user input of adding a deadline task to LIST
+     * @param userInput for user input
+     * @return chatbot's response to succesfully adding task
+     */
+    public static String respondTo(String userInput) {
+        String[] parts = userInput.split("/by", 2);
+        String description = parts[0].replaceFirst("deadline", "").trim();
+        String by = parts[1].trim();
+
+        LocalDate byDate = isDate(by);
+        Deadline newDeadline = new Deadline(description, byDate);
+        Constants.LIST.add(newDeadline);
+
+        return (Constants.ADDTASK
+                + newDeadline + "\n"
+                + Helper.tasksLeft(Constants.LIST.size()));
+    }
 
 }
