+++ src/main/java/tasks/Event.java
@@ -1,7 +1,12 @@
-package Tasks;
+package tasks;
+
+import static resources.DateHandler.isDate;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
+
+import resources.Constants;
+import resources.Helper;
 
 /**
  * Event class that extends Task
@@ -40,5 +45,28 @@ public class Event extends Task {
 
     public LocalDate getTo() {
         return this.to;
+    }
+
+    /**
+     * respond to user input of adding an event task to LIST
+     * @param userInput for user input
+     * @return chatbot's respond to adding an event task to LIST
+     */
+    public static String respondTo(String userInput) {
+        String[] fromSplit = userInput.split("/from", 2);
+        String description = fromSplit[0].replaceFirst("event", "").trim();
+
+        String[] toSplit = fromSplit[1].split("/to", 2);
+        String from = toSplit[0].trim();
+        String to = toSplit[1].trim();
+
+        LocalDate fromDate = isDate(from);
+        LocalDate toDate = isDate(to);
+        Event newEvent = new Event(description, fromDate, toDate);
+        Constants.LIST.add(newEvent);
+
+        return (Constants.ADDTASK
+                + newEvent + "\n"
+                + Helper.tasksLeft(Constants.LIST.size()));
     }
 }
